Changed Namespace of ProbData. New version of Reader.

+++ Scip/scip-3.0.1/examples/CallableLibrary/src/GomoryHuTree.h
@@ -38,7 +38,7 @@ using namespace std;
 
 typedef struct GraphNode
 {
-	int 				  stadtid;
+	long int 			  stadtid;
 	string 				  name;
 	int 				  kreisid;
 	int     			  bewohner;
